+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/SpotlessCheckMojo.java
@@ -27,6 +27,7 @@ import org.apache.maven.plugins.annotations.Mojo;
 import com.diffplug.spotless.Formatter;
 import com.diffplug.spotless.PaddedCell;
 import com.diffplug.spotless.extra.integration.DiffMessageFormatter;
+import com.diffplug.spotless.maven.incremental.UpToDateChecker;
 
 /**
  * Performs code formatting analysis and prints all violations to the console.
@@ -36,13 +37,22 @@ import com.diffplug.spotless.extra.integration.DiffMessageFormatter;
 public class SpotlessCheckMojo extends AbstractSpotlessMojo {
 
 	@Override
-	protected void process(Iterable<File> files, Formatter formatter) throws MojoExecutionException {
+	protected void process(Iterable<File> files, Formatter formatter, UpToDateChecker upToDateChecker) throws MojoExecutionException {
 		List<File> problemFiles = new ArrayList<>();
 		for (File file : files) {
+			if (upToDateChecker.isUpToDate(file.toPath())) {
+				if (getLog().isDebugEnabled()) {
+					getLog().debug("Spotless will not check an up-to-date file: " + file);
+				}
+				continue;
+			}
+
 			try {
 				PaddedCell.DirtyState dirtyState = PaddedCell.calculateDirtyState(formatter, file);
 				if (!dirtyState.isClean() && !dirtyState.didNotConverge()) {
 					problemFiles.add(file);
+				} else {
+					upToDateChecker.setUpToDate(file.toPath());
 				}
 			} catch (IOException e) {
 				throw new MojoExecutionException("Unable to format file " + file, e);